requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-dotenv
requests

